Move all classes to respected packages
