+++ src/main/java/com/codecool/java/expertsystem/Main.java
@@ -1,5 +1,8 @@
 package com.codecool.java.expertsystem;
 
+import com.codecool.java.expertsystem.parser.FactParser;
+import com.codecool.java.expertsystem.parser.RuleParser;
+
 public class Main {
 
     public static void main(String[] args) {
