+++ src/main/java/com/codecool/java/expertsystem/model/SingleValue.java
@@ -1,4 +1,4 @@
-package com.codecool.java.expertsystem;
+package com.codecool.java.expertsystem.model;
 
 import java.util.ArrayList;
 import java.util.List;
